finished Error + RT analyses
For Error, the inverse transform was not feasible, and had to be
replaced with Yeo-Johnson. The results do not change. A simple approach
to model building is displayed.

For RT, there are no correlational effects, and no model is built (nor
is it clear how one ought to do this).